requirements.txt
Flask
torch
Pillow
ultralytics
